cmd/derpprobe: don't alert for smaller failures.
There is a Cosmic Background level of DERP Unreachability,
with individual nodes or regions becoming unreachable briefly
and returning a short time later.

Returning a 500 error pages a human. Being awoken at 3am for
a transient error is annoying.

For relatively small levels of badness, don't page a human.
Just post to Slack. We have enough experience with how often
this happens to think that it won't drown the Slack channel
in updates.

Signed-off-by: Denton Gentry <dgentry@tailscale.com>